revert: undo PROT-019 file split; harden PROT-015 audit-repro test
Folds ControllerImpl.hpp back into Controller.cpp.  The split was a
cosmetic reorganization, not a behavioral fix, and the accompanying
PROT-019 audit-repro test only checked Controller.cpp's line count — a
self-imposed 600-line threshold that nothing in the project enforces.

Removes:
  - src/rogue/protocols/rssi/ControllerImpl.hpp (textually included from
    Controller.cpp; no header guard, no separate compilation, no value)
  - tests/cpp/protocols/rssi/test_rssi_monolith_audit_repro.cpp (line-
    count tautology that only validated the split it shipped with)

Strengthens the PROT-015 busy-loop audit-repro test: the previous version
scanned the whole file for any 'notify_all' or 'wait_for' substring, both
of which already appear on origin/main in unrelated paths (runThread,
transportRx).  The test passed even with the usleep(10) busy-loop back in
applicationRx().  The new test asserts 'usleep(' is absent from
Controller.cpp — strict, file-wide, and unambiguous since the audit
removed the only usleep() call from this translation unit.

Reproducibility verified locally:
  - PROT-002 (Controller::start RAII thread): un-patch -> ctest fails;
    restore -> ctest passes.
  - PROT-015 (applicationRx busy-loop -> condvar): un-patch -> ctest
    fails under the strengthened test; restore -> ctest passes.

C++ test suite: 47/48 pass (the api-smoke env-shadowing failure is
pre-existing and documented in the PR description).
