+++ tests/cpp/protocols/rssi/CMakeLists.txt
@@ -19,13 +19,3 @@ rogue_add_cpp_test(rogue-cpp-protocols-rssi-busyloop-audit-repro
 )
 target_compile_definitions(rogue-cpp-protocols-rssi-busyloop-audit-repro
    PRIVATE ROGUE_SRC_DIR="${CMAKE_SOURCE_DIR}")
-
-rogue_add_cpp_test(rogue-cpp-protocols-rssi-monolith-audit-repro
-   SOURCES
-      test_rssi_monolith_audit_repro.cpp
-   LABELS
-      cpp-core
-      no-python
-)
-target_compile_definitions(rogue-cpp-protocols-rssi-monolith-audit-repro
-   PRIVATE ROGUE_SRC_DIR="${CMAKE_SOURCE_DIR}")
